+++ CSETWebApi/CSETWeb_Api/CSETWebCore.Business/Assessment/AssessmentBusiness.cs
@@ -33,6 +33,7 @@ namespace CSETWebCore.Business.Assessment
         private readonly IContactBusiness _contactBusiness;
         private readonly ISalBusiness _salBusiness;
         private readonly IMaturityBusiness _maturityBusiness;
+        private readonly IACETMaturityBusiness _acetMaturityBusiness;
         private readonly IAssessmentUtil _assessmentUtil;
         private readonly IStandardsBusiness _standardsBusiness;
         private readonly IDiagramManager _diagramManager;
