+++ CSETWebApi/CSETWeb_Api/CSETWeb_ApiCore/Controllers/HomeACETController.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace CSETWebCore.Api.Controllers
+{
+    public class HomeACETController : Controller
+    {
+        public IActionResult Index()
+        {
+            return View();
+        }
+    }
+}
