+++ CSETWebApi/CSETWeb_Api/CSETWeb_ApiCore/Controllers/MaturityAcetController.cs
@@ -1,0 +1,130 @@
+﻿//////////////////////////////// 
+// 
+//   Copyright 2024 Battelle Energy Alliance, LLC  
+// 
+// 
+//////////////////////////////// 
+using CSETWebCore.Business.Acet;
+using CSETWebCore.Business.Maturity;
+using CSETWebCore.DataLayer.Model;
+using CSETWebCore.Interfaces.AdminTab;
+using CSETWebCore.Interfaces.Helpers;
+using CSETWebCore.Interfaces.Reports;
+using Microsoft.AspNetCore.Mvc;
+
+
+namespace CSETWebCore.Api.Controllers
+{
+    public class MaturityAcetController : ControllerBase
+    {
+        private readonly ITokenManager _tokenManager;
+        private readonly CSETContext _context;
+        private readonly IAssessmentUtil _assessmentUtil;
+        private readonly IAdminTabBusiness _adminTabBusiness;
+        private readonly IReportsDataBusiness _reports;
+
+
+        /// <summary>
+        /// CTOR
+        /// </summary>
+        public MaturityAcetController(ITokenManager tokenManager, CSETContext context, IAssessmentUtil assessmentUtil,
+            IAdminTabBusiness adminTabBusiness, IReportsDataBusiness reports)
+        {
+            _tokenManager = tokenManager;
+            _context = context;
+            _assessmentUtil = assessmentUtil;
+            _adminTabBusiness = adminTabBusiness;
+            _reports = reports;
+        }
+
+
+        /// <summary>
+        /// Get maturity calculations
+        /// </summary>
+        /// <returns></returns>
+        [HttpGet]
+        [Route("api/getMaturityResults")]
+        public IActionResult GetMaturityResults()
+        {
+            int assessmentId = _tokenManager.AssessmentForUser();
+            var lang = _tokenManager.GetCurrentLanguage();
+
+            ACETMaturityBusiness manager = new ACETMaturityBusiness(_context, _assessmentUtil, _adminTabBusiness);
+            var maturity = manager.GetMaturityAnswers(assessmentId, lang);
+
+            return Ok(maturity);
+        }
+
+
+        /// <summary>
+        /// Get maturity calculations
+        /// </summary>
+        /// <returns></returns>
+        [HttpGet]
+        [Route("api/getIseMaturityResults")]
+        public IActionResult GetIseMaturityResults()
+        {
+            int assessmentId = _tokenManager.AssessmentForUser();
+
+            var manager = new ACETMaturityBusiness(_context, _assessmentUtil, _adminTabBusiness);
+            var maturity = manager.GetIseMaturityAnswers(assessmentId);
+
+            return Ok(maturity);
+        }
+
+
+        /// <summary>
+        /// Get maturity range based on IRP rating
+        /// </summary>
+        /// <returns></returns>
+        [HttpGet]
+        [Route("api/getMaturityRange")]
+        public IActionResult GetMaturityRange()
+        {
+            int assessmentId = _tokenManager.AssessmentForUser();
+            ACETMaturityBusiness manager = new ACETMaturityBusiness(_context, _assessmentUtil, _adminTabBusiness);
+            var maturityRange = manager.GetMaturityRange(assessmentId);
+            return Ok(maturityRange);
+        }
+
+
+        /// <summary>
+        /// Get IRP total for maturity
+        /// </summary>
+        /// <returns></returns>
+        [HttpGet]
+        [Route("api/getOverallIrpForMaturity")]
+        public IActionResult GetOverallIrp()
+        {
+            int assessmentId = _tokenManager.AssessmentForUser();
+            return Ok(new AcetBusiness(_context, _assessmentUtil, _adminTabBusiness).GetOverallIrp(assessmentId));
+        }
+
+
+        /// <summary>
+        /// Get target band for maturity
+        /// </summary>
+        /// <returns></returns>
+        [HttpGet]
+        [Route("api/getTargetBand")]
+        public IActionResult GetTargetBand()
+        {
+            int assessmentId = _tokenManager.AssessmentForUser();
+            return Ok(new ACETMaturityBusiness(_context, _assessmentUtil, _adminTabBusiness).GetTargetBandOnly(assessmentId));
+        }
+
+
+        /// <summary>
+        /// Set target band for maturity rating
+        /// </summary>
+        /// <returns></returns>
+        [HttpPost]
+        [Route("api/setTargetBand")]
+        public IActionResult SetTargetBand([FromBody] bool value)
+        {
+            int assessmentId = _tokenManager.AssessmentForUser();
+            new ACETMaturityBusiness(_context, _assessmentUtil, _adminTabBusiness).SetTargetBandOnly(assessmentId, value);
+            return Ok();
+        }
+    }
+}
